# gravitee-rest-api-portal/gravitee-rest-api-portal-rest/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Gravitee.io Portal Rest API
  description: API dedicated to the devportal part of Gravitee
  termsOfService: terms
  contact:
    name: The Gravitee Team
    url: https://gravitee.io
    email: contact@graviteesource.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: "3.0.0"
servers:
  - url: http://demo.gravitee.io/portal/{envId}
    description: The portal API for a given environment
    variables:
      envId:
        default: DEFAULT
        description: ID of the environment
...
security:
  - BasicAuth: []
  - CookieAuth: []
paths:
  /apis:
    get:
      tags:
        - Api
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/apiContextPathQueryParam'
        - $ref: '#/components/parameters/apiLabelQueryParam'
        - $ref: '#/components/parameters/apiVersionQueryParam'
        - $ref: '#/components/parameters/apiNameQueryParam'
        - $ref: '#/components/parameters/apiViewQueryParam'
        - $ref: '#/components/parameters/apiCategoryQueryParam'
        - $ref: '#/components/parameters/apiCategoryExcludedQueryParam'
      summary: List APIs
      description: |
        List public APIs for anonymous requests.
        List all the APIs the current user is allowed to view for authenticated requests.

        The list can be filtered according to query parameters.

        By default the list is sorted by alphabetic order.
        If a **cat** query param is sent, a specific sorting can be applied. Please refer to **cat** description for detail.
      operationId: getApis
      security: []
      responses:
        200:
          description: List accessible APIs for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApisResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
  /apis/_search:
    post:
      tags:
        - Api
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/apiQueryParam'
      summary: Search APIs with a query
      description: |
        Same as `/apis` but with a search query as a parameter. Search for API using the search engine.
        Supports pagination.
      operationId: searchApis
      security: []
      responses:
        200:
          description: Search accessible APIs for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApisResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
  /apis/{apiId}:
    parameters:
      - $ref: '#/components/parameters/apiIdParam'
    get:
      tags:
        - Api
      parameters:
        - $ref: '#/components/parameters/apiIncludeParam'
      summary: Get the API definition
      description:  |
        Get the detail of an API.

        This API has to be accessible by the current user, otherwise a 404 will be returned.
      operationId: getApiByApiId
      security: []
      responses:
        200:
          description: API definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api'
        404:
          $ref: '#/components/responses/APINotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /subscriptions:
    get:
      tags:
        - Subscription
      parameters:
        - $ref: '#/components/parameters/apiIdQueryParam'
        - $ref: '#/components/parameters/applicationIdQueryParam'
        - $ref: '#/components/parameters/subscriptionStatusesQueryParam'
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
      description: |
        List all ACCEPTED, PAUSED & PENDING subscriptions, filtered by api and/or by application. At least an api or an application must be provided.

        User must have the APPLICATION_SUBSCRIPTION[READ] permission to list subscription with application query param.\
        User must have the API_SUBSCRIPTION[READ] permission to list subscription with api query param.

      operationId: getSubscriptions
      responses:
        200:
          description: List of Subscriptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionsResponse'
        400:
          description: At least an api or an application must be provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          $ref: '#/components/responses/PermissionError'
        500:
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Subscription
      requestBody:
        description: Use to create a subscription.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionInput'
      summary: Create a subscription.
      description: |
        Create a new subscription.

        User must have APPLICATION_SUBSCRIPTION[CREATE] permission.
      operationId: createSubscription
      responses:
        200:
          description: Created subscription.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        400:
          description: Input must not be null.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          $ref: '#/components/responses/PermissionError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /subscriptions/{subscriptionId}:
    parameters:
      - $ref: '#/components/parameters/subscriptionIdParam'
    get:
      tags:
        - Subscription
      parameters:
        - $ref: '#/components/parameters/subscriptionIncludeQueryParam'

      summary: Get a subscription.
      description: |
        Get a subscription.

        User must have API_SUBSCRIPTION[CREATE] or APPLICATION_SUBSCRIPTION[CREATE] permission.
      operationId: getSubscriptionById
      responses:
        200:
          description: One subscription.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  parameters:
...
    apiIdParam:
      name: apiId
      in: path
      required: true
      description: Id of an API.
      schema:
        type: string
...
    subscriptionIdParam:
      name: subscriptionId
      in: path
      required: true
      description: Id of a subscription.
      schema:
        type: string
...
    apiIdQueryParam:
      name: apiId
      in: query
      required: false
      description: Id of an api.
      schema:
        type: string
    applicationIdQueryParam:
      name: applicationId
      in: query
      required: false
      description: Id of an application.
      schema:
        type: string
...
    pageNumberParam:
      name: page
      in: query
      required: false
      description: The page number for pagination.
      schema:
        type: integer
        default: 1
    pageSizeParam:
      name: size
      in: query
      required: false
      description: |
        The number of items per page for pagination.
        If the size is 0, the response contains only metadata and returns the values as for a non-paged resource.
        If the size is -1, the response contains all datas.
      schema:
        type: integer
        default: 10

    #Apis filtering
    apiContextPathQueryParam:
      name: context-path
      in: query
      required: false
      description: The context-path of an API.
      schema:
        type: string
    apiLabelQueryParam:
      name: label
      in: query
      required: false
      description: One of the labels of an API.
      schema:
        type: string
    apiVersionQueryParam:
      name: version
      in: query
      required: false
      description: The version of an API.
      schema:
        type: string
    apiNameQueryParam:
      name: name
      in: query
      required: false
      description: The name of an API.
      schema:
        type: string
    apiViewQueryParam:
      name: view
      in: query
      required: false
      description: One of the view linked with an API.
      schema:
        type: string
    apiCategoryQueryParam:
      name: cat
      in: query
      required: false
      description: |
        Name of a CATEGORY. A category is a static filter used to pre-select some APIs. It can be combined with others query params.
          - FEATURED : top APIs. Sorted by alphabetical order.
          - MINE : APIs subscribed by the current user. Sorted by alphabetical order.
          - STARRED: Rated APIs. Sorted by Ratings DESC, nb of ratings DESC, alphabetic order.
          - TRENDINGS: Subscribed APIs. Sorted by number of subscriptions DESC, alphabetic order.
      schema:
        $ref: '#/components/schemas/CategoryApiQuery'
    apiCategoryExcludedQueryParam:
      name: -cat
      in: query
      required: false
      description: |
        Name of a CATEGORY to exclude. Excluded category has priority over category. A category is a static filter used to pre-select some APIs. It can be combined with others query params.
          - FEATURED : not in top APIs. Sorted by alphabetical order.
          - MINE : APIs not subscribed by the current user. Sorted by alphabetical order.
          - STARRED: not rated APIs. Sorted by Ratings DESC, nb of ratings DESC, alphabetic order.
          - TRENDINGS: not subscribed APIs. Sorted by number of subscriptions DESC, alphabetic order.
      schema:
        $ref: '#/components/schemas/CategoryApiQuery'
    apiQueryParam:
      name: q
      in: query
      required: true
      description: query string to be used in the search engine
      schema:
        type: string
    apiIncludeParam:
      name: include
      in: query
      required: false
      description: Comma-separated list of related objects to include in the response.
      schema:
        type: array
        items:
          type: string
          enum:
            - pages
            - plans
...
    subscriptionStatusesQueryParam:
      name: statuses
      in: query
      required: false
      description: Id of an application.
      schema:
        type: array
        items:
          type: string
          enum:
            - ACCEPTED
            - CLOSED
            - PAUSED
            - PENDING
            - REJECTED
    subscriptionIncludeQueryParam:
      name: include
      in: query
      required: false
      description: Comma-separated list of related objects to include in the response.
      schema:
        type: array
        items:
          type: string
          enum:
            - keys
    
    #Users filtering
...
  schemas:
...
    ErrorResponse:
      properties:
        errors :
          description: List of errors
          type: array
          items:
            $ref: '#/components/schemas/Error'
    ApisResponse:
      properties:
        data:
          description: List of API.
          type: array
          items:
            $ref: '#/components/schemas/Api'
        metadata:
            $ref: '#/components/schemas/MetadataMap'
        links:
          $ref: '#/components/schemas/Links'
...
    SubscriptionsResponse:
      properties:
        data:
          description: List of subscriptions.
          type: array
          items:
            $ref: '#/components/schemas/Subscription'
        metadata:
            $ref: '#/components/schemas/MetadataMap'
        links:
          $ref: '#/components/schemas/Links'
...
    Api:
      description: |
        Describes an API.
      required:
        - id
        - name
        - version
        - description
        - owner
      properties:
        id:
          description: Unique identifier of an API.
          type: string
        name:
          description: Name of the API.
          type: string
        version:
          description: Version of the API.
          type: string
        description:
          description: Description of the API.
          type: string
        draft:
          type: boolean
          description: Whether or not the API is in draft.
        public:
          type: boolean
          description: Whether or not the API is public.
        running:
          type: boolean
          description: Whether or not the API is running.
        entrypoints:
          description: List of all the avaible endpoints to call the API.
          type: array
          items:
            type: string
        labels:
          description: List of labels linked to this API.
          type: array
          items:
            type: string
        owner:
          $ref: '#/components/schemas/User'
        updated_at:
          description: Last update date and time.
          type: string
          format: date-time
        views:
          description: List of views this API belongs to.
          type: array
          items:
            $ref: '#/components/schemas/View'
        rating_summary:
          $ref: '#/components/schemas/RatingSummary'
        _links:
          $ref: '#/components/schemas/ApiLinks'

        ###################
        # may be included #
        ###################
        pages:
          description: |
            Only returned with */apis/{apiId}*.
            Need *include* query param to contain 'pages'.

            The documentation pages of this API. Same as */apis/{apiId}/pages*.
          type: array
          items:
            $ref: '#/components/schemas/Page'
        plans:
          description: |
            Only returned with */apis/{apiId}*.
            Need *include* query param to contain 'plans'.

            The plans of this API. Same as */apis/{apiId}/plans*.
          type: array
          items:
            $ref: '#/components/schemas/Plan'
...
    Page:
      required:
        - id
        - name
        - type
        - order
      properties:
        id:
          description: Unique identifier of a page.
          type: string
        name:
          description: Name of the page.
          type: string
        type:
          description: Type of documentation.
          type: string
          enum:
          - SWAGGER
          - MARKDOWN
          - FOLDER
          - ROOT
          - LINK
        order:
          description: Order of the documentation page in its folder.
          type: integer
        parent:
          description: Parent page. MAY be null.
          type: string
        updated_at:
          description: Last update date and time.
          type: string
          format: date-time
        configuration:
          $ref: '#/components/schemas/PageConfiguration'
        metadata:
          description: Array of metadata about the page. This array is filled when the page has been fetched from a distant source (GitHub, GitLab, etc...).
          type: array
          items:
            $ref: '#/components/schemas/Metadata'
        _links:
          $ref: '#/components/schemas/PageLinks'

        ###################
        # may be included #
        ###################
        content:
          description: |
            Only returned with */apis/{apiId}/pages/{pageId}* and */pages/{pageId}*.
            Need *include* query param to contain 'content'.

            The content of the page.
          type: string
    Plan:
      required:
        - id
        - name
        - security
        - type
        - description
        - validation
        - order
        - comment_required
      properties:
        id:
          description: Unique identifier of a plan.
          type: string
        name:
          description: Name of the plan.
          type: string
        security:
          description: Security used with this plan.
          type: string
          enum:
            - API_KEY
            - KEY_LESS
            - JWT
            - OAUTH2
        description:
          description: Description of the plan.
          type: string
        characteristics:
          description: List of additionnal terms to describe the plan.
          type: array
          items:
            type: string
        validation:
          description: Type of validation for subscription requests.
          type: string
          enum:
            - AUTO
            - MANUAL
        order:
          description: Priority order
          type: integer
        subscribed:
          description: True if at least one application linked to the current user has subscribed to this plan.
          type: boolean
        comment_required:
          description: True if a comment is required when a subscription is created.
          type: boolean
        comment_question:
          description: Content of the message sent to a user creating a subscription.
          type: string
...
    Subscription:
      required:
        - id
        - api
        - application
        - plan
        - status
      properties:
        id:
          description: Unique identifier of a subscription.
          type: string
        api:
          description: Subscribed API.
          type: string
        application:
          description: Subscribing application.
          type: string
        plan:
          description: Subscribed plan.
          type: string
        request:
          description: Comment of the user when asking for a subscription.
          type: string
        reason:
          description: Reason of the user when processing (accepting/rejecting) a subscription.
          type: string
        created_at:
          description: Creation date and time of the subscription.
          type: string
          format: date-time
        processed_at:
          description: Date and time when the subscription request was processed.
          type: string
          format: date-time
        start_at:
          description: Start date and time of the subscription.
          type: string
          format: date-time
        end_at:
          description: Expiration date and time of the subscription.
          type: string
          format: date-time
        paused_at:
          description: Paused date and time of the subscription.
          type: string
          format: date-time
        closed_at:
          description: Closed date and time of the subscription.
          type: string
          format: date-time
        subscribed_by:
          description: The user who subscribed.
          type: string
        status:
          description: Status of the subscription.
          type: string
          enum:
            - PENDING
            - ACCEPTED
            - CLOSED
            - REJECTED
            - PAUSED
        ###################
        # may be included #
        ###################
        keys:
          description: |
            Only returned with */subscriptions/{subscriptionId}*.
            Need *include* query param to contain 'keys'.

            List of APIKeys of the subscription.
          type: array
          items:
            $ref: '#/components/schemas/Key'
    User:
      properties:
        id:
          description: Unique identifier of a user.
          type: string
        reference:
          description: Unique reference if user comes from external source. Use for search only.
          type: string
        first_name:
          type: string
        last_name:
          type: string
        display_name:
          type: string
        email:
          type: string
        config:
          $ref: '#/components/schemas/UserConfig'
        _links:
          $ref: '#/components/schemas/UserLinks'
...
    View:
      properties:
        id:
          description: Unique identifier of a view.
          type: string
        name:
          type: string
        description:
          type: string
        default_view:
          type: boolean
        order:
          type: integer
        total_apis:
          type: integer
          format: int64
        _links:
          $ref: '#/components/schemas/ViewLinks'
...
    SubscriptionInput:
      properties:
        application:
          description: Id of the application which subscribe
          type: string
        plan:
          description: Id of the plan to which the application subscribe
          type: string
        request:
          description: A request message to the api owner why a user want to subscribe
          type: string
...
    RatingSummary:
      properties:
        average:
          type: number
          format: double
        count:
          type: number
        one_star:
          type: number
          format: double
        two_star:
          type: number
          format: double
        three_star:
          type: number
          format: double
        four_star:
          type: number
          format: double
        five_star:
          type: number
          format: double
...
    MetadataMap:
      description: Map of Map of Object
      type: object
      additionalProperties:
        type: object
        additionalProperties:
          type: object
...
    CategoryApiQuery:
      type: string
      enum:
        - FEATURED
        - MINE
        - STARRED
        - TRENDINGS
...
    Links:
      properties:
        self:
          type: string
        first:
          type: string
        last:
          type: string
        prev:
          type: string
        next:
          type: string
    ApiLinks:
      properties:
        self:
          type: string
        links:
          type: string
        metrics:
          type: string
        pages:
          type: string
        picture:
          type: string
        plans:
          type: string
        ratings:
          type: string
...
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PermissionError:
      description: Insufficient rights to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    APINotFoundError:
      description: API not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    CookieAuth:
      type: apiKey
      in: cookie
      name: Auth-Graviteeio-APIM